pyproject.toml
[tool.poetry]
name = "summarization"
version = "0.0.1"
description = "summarization dashboard"
authors = ["Jumin, Aditya, Sridhar <ffpgmaas@gmail.com>"]
repository = "https://github.com/fpgmaas/summarization"
documentation = "https://fpgmaas.github.io/summarization/"
readme = "README.md"
packages = [
  {include = "summarization"}
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pymongo = "^4.6.1"
mysql-connector-python = "^8.2.0"
kafka-python = "^2.0.2"
requests = "^2.31.0"
pandas = "^2.1.3"
sqlalchemy = "2.0.18"
uvicorn = "^0.24.0.post1"
fastapi = {extras = ["all"], version = "^0.104.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["summarization"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]



